refactor: move grab rail handles into their own partial
