+++ src/CtrDxEditor.Shared/Rendering/LevelCanvas.Input.cs
@@ -82,24 +82,6 @@ namespace CtrDxEditor.Rendering
                     HitTolerance(9));
         }
 
-        /// <summary>What part of the selected movable grab's rail a level point is over, or <see cref="GrabRail.Handle.None"/>.</summary>
-        /// <remarks>
-        /// The hit-testing itself lives in <see cref="GrabRail"/>; here we only supply the selected grab's geometry and
-        /// the screen-derived tolerances: ~9 px for the end caps, the hook's own footprint, and the bar's half thickness.
-        /// </remarks>
-        /// <param name="levelPt">The point to test, in level coordinates.</param>
-        /// <returns>The rail handle under the point, or <see cref="GrabRail.Handle.None"/>.</returns>
-        private GrabRail.Handle HitRail(Vec2 levelPt)
-        {
-            return IsSingleSelection
-                && SelectedObject is { Type: "grab" } sel
-                && View.Zoom > 0
-                && GrabRenderer.DrawsMovableRail(sel)
-                && GrabRail.Of(sel) is { } g
-                ? GrabRail.HitTest(g, levelPt, endTolerance: HitTolerance(9), hookTolerance: 24, barThickness: 20)
-                : GrabRail.Handle.None;
-        }
-
         /// <summary>What part of the selected spike/bouncer resize affordance a level point is over.</summary>
         /// <param name="levelPt">The point to test, in level coordinates.</param>
         /// <returns>The strip resize handle under the point, or <see cref="SpikeResize.Handle.None"/>.</returns>
@@ -285,70 +267,12 @@ namespace CtrDxEditor.Rendering
             RotationDialTargetResolver.ApplyAngle(obj, target, angle, center);
         }
 
-        /// <summary>
-        /// Applies the active rail drag to the grab: sliding moves the hook (object x/y) and its offset together so the
-        /// rail stays put; resizing an end rewrites <c>moveLength</c> (and <c>moveOffset</c> for the near end). All
-        /// constrained by <see cref="GrabRail"/> so the hook never leaves the rail.
-        /// </summary>
-        /// <param name="grab">The grab object being edited.</param>
-        /// <param name="g">The grab's current rail geometry.</param>
-        /// <param name="levelPt">The pointer position in level coordinates.</param>
-        private void ApplyRailDrag(LevelObject grab, GrabRail.Geometry g, Vec2 levelPt)
-        {
-            switch (_railDrag)
-            {
-                case GrabRail.Handle.SlideHook:
-                    (double hookAxis, double offset) = GrabRail.SlideHook(g, levelPt);
-                    if (g.Vertical)
-                    {
-                        grab.Y = (int)Math.Round(hookAxis);
-                    }
-                    else
-                    {
-                        grab.X = (int)Math.Round(hookAxis);
-                    }
-                    grab.SetAttr("moveOffset", Whole(offset));
-                    break;
-                case GrabRail.Handle.ResizeEnd:
-                    grab.SetAttr("moveLength", Whole(GrabRail.ResizeEnd(g, levelPt)));
-                    break;
-                case GrabRail.Handle.ResizeStart:
-                    (double offA, double length) = GrabRail.ResizeStart(g, levelPt);
-                    grab.SetAttr("moveOffset", Whole(offA));
-                    grab.SetAttr("moveLength", Whole(length));
-                    break;
-                case GrabRail.Handle.MoveBar:
-                case GrabRail.Handle.None:
-                default:
-                    break;
-            }
-        }
-
         /// <summary>Rounds a level-space value to a whole number and formats it with the invariant culture for an attribute.</summary>
         /// <param name="value">The value to round and format.</param>
         /// <returns>The rounded integer as an invariant-culture string.</returns>
         private static string Whole(double value)
         {
             return ((int)Math.Round(value)).ToString(CultureInfo.InvariantCulture);
-        }
-
-        /// <summary>
-        /// The cursor for a rail handle: a horizontal rail end/hook reads as a horizontal resize, a vertical one as a
-        /// vertical resize (the hook slides along the same axis). The bar keeps the default arrow — it is still
-        /// draggable to move the whole grab, but a move cursor over the whole rail is noisy.
-        /// </summary>
-        /// <param name="handle">The rail handle under the cursor.</param>
-        /// <returns>The cursor to display for that handle.</returns>
-        private Cursor CursorForHandle(GrabRail.Handle handle)
-        {
-            return handle switch
-            {
-                GrabRail.Handle.ResizeStart or GrabRail.Handle.ResizeEnd or GrabRail.Handle.SlideHook =>
-                    SelectedObject is { } s && GrabRail.Vertical(s) ? VResizeCursor : ResizeCursor,
-                GrabRail.Handle.MoveBar => Cursor.Default,
-                GrabRail.Handle.None => Cursor.Default,
-                _ => Cursor.Default,
-            };
         }
 
         /// <summary>Cursor for a strip resize handle based on the object's current rotation.</summary>
